Move cert/HTTP-related dynamic waches to cert/http package
